+++ Objt/NonRollable.java
@@ -1,7 +1,7 @@
 package Objt;
 
 abstract class NonRollable extends Objects {
-    NonRollable(float mass, float mu, float[] s0, float[] u, int objectID,  int objectTypeID, float point_collision)
+    NonRollable(double mass, double mu, double[] s0, double[] u, int objectID,  int objectTypeID, double point_collision)
     {
         super(mass, mu, s0,u,objectID, objectTypeID, point_collision);
     }
