+++ Objt/Sphere.java
@@ -1,9 +1,9 @@
 package Objt;
 
 public class Sphere extends Rollable {
-    private float radius;
+    private double radius;
 
-    public Sphere(float mass, float mu, float[] s0, float[] u, int objectID, int objectTypeID, float radius) 
+    public Sphere(double mass, double mu, double[] s0, double[] u, int objectID, int objectTypeID, double radius) 
     {
         super(mass, mu, s0, u, objectID, objectTypeID,radius/2);
         this.radius = radius;
@@ -11,11 +11,11 @@ public class Sphere extends Rollable {
 
 
 
-    public float getRadius() {
+    public double getRadius() {
         return radius;
     }
 
-    public void setRadius(float radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
